+++ app/Filament/Widgets/ProductosWidget.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Filament\Widgets;
+
+use Filament\Widgets\Widget;
+
+class ProductosWidget extends Widget
+{
+    protected static ?int $sort = 4;
+
+    protected string $view = 'filament.widgets.productos-widget';
+}